requirements.txt
Flask==3.0.2
Flask-SQLAlchemy==3.1.1
python-dotenv==1.0.1
flask-cors==4.0.0
Flask-Migrate==4.0.5  # For database migrations 
Flask-Login==0.6.3
Werkzeug==3.0.1